impr: extract signature module
Extract the signature module and related modules (param and region) into a
separate crate.
