+++ .travis.yml
@@ -46,6 +46,7 @@ script:
   - echo "Running Rusoto tests B" && travis_wait 40 travis-cargo test -- --lib --features packageb -v
   - echo "Running Rusoto tests" && cd rusoto && travis_wait 40 cargo test --lib --all -v && cd ..
   - echo "Running rusoto_credential tests" && cd credential && travis_wait 40 travis-cargo test && cd ..
+  - echo "Running rusoto_signature tests" && cd signature && travis_wait 40 travis-cargo test && cd ..
   - |
       echo "Running cargo docs on stable Rust on Linux" &&
       if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then
