--- rusoto/core/test_resources/multiple_profile_credentials
@@ -1,7 +1,0 @@
-[foo]
-aws_access_key_id = foo_access_key
-aws_secret_access_key = foo_secret_key
-
-[bar]
-aws_access_key_id = bar_access_key
-aws_secret_access_key = bar_secret_key
