+++ signature/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+authors = [
+    "Anthony DiMarco <ocramida@gmail.com>",
+    "Jimmy Cuadra <jimmy@jimmycuadra.com>",
+    "Matthew Mayer <matthewkmayer@gmail.com>",
+    "Nikita Pekin <contact@nikitapek.in>"
+]
+description = "AWS signature tooling"
+keywords = ["AWS", "Amazon"]
+license = "MIT"
+name = "rusoto_signature"
+readme = "README.md"
+repository = "https://github.com/rusoto/rusoto"
+version = "0.1.0"
+homepage = "https://www.rusoto.org/"
+
+[dependencies]
+log = "0.3.6"
+ring = "0.7"
+rustc-serialize = "0.3.21"
+time = "0.1.35"
+url = "1.2.0"
+
+[dependencies.rusoto_credential]
+path = "../credential"
+version = "0.6.0"
+
+[dependencies.clippy]
+optional = true
+version = "0.0"
+
+[features]
+nightly-testing = ["clippy"]
