+++ signature/README.md
@@ -1,0 +1,12 @@
+# Rusoto signature
+
+This crate contains the functionality necessary for signing AWS API requests. It
+was designed for use in [rusoto][rusoto].
+
+## Development
+
+The Rusoto signature crate is packaged and published separately from the Rusoto
+crate itself. The crate does not depend on Rusoto itself. The signature crate
+is published as a versioned dependency of Rusoto on crates.io.
+
+[rusoto]: https://github.com/rusoto/rusoto
