+++ signature/test_resources/multiple_profile_credentials
